refactor: move `AxesConfiguration` from `erlab.analysis.kspace` to `erlab.constants`
